\name{coerce}
\alias{coerce,CXCursor,CXSourceRange-method}
\alias{coerce,CXCursor,character-method}
\alias{coerce,CXCursor,list-method}
\alias{coerce,CXDiagnostic,character-method}
\alias{coerce,CXFile,character-method}
\alias{coerce,CXTranslationUnit,character-method}
\alias{coerce,CXType,character-method}
\alias{coerce,FunctionDecl,CXSourceRange-method}
\alias{coerce,FunctionDecl,character-method}
\alias{coerce,NativeElement,CXCursor-method}
\alias{coerce,RCallFunctionTable,character-method}
\alias{coerce,RCallableFunctions,RegistrationTable-method}
\alias{coerce,RDotCFunctionList,character-method}
\alias{coerce,character,CXAvailabilityKind-method}
\alias{coerce,character,CXBinaryOperatorKind-method}
\alias{coerce,character,CXCallingConv-method}
\alias{coerce,character,CXChildVisitResult-method}
\alias{coerce,character,CXChoice-method}
\alias{coerce,character,CXCodeComplete_Flags-method}
\alias{coerce,character,CXCompletionChunkKind-method}
\alias{coerce,character,CXCompletionContext-method}
\alias{coerce,character,CXCursorKind-method}
\alias{coerce,character,CXCursor_ExceptionSpecificationKind-method}
\alias{coerce,character,CXDiagnosticDisplayOptions-method}
\alias{coerce,character,CXDiagnosticSeverity-method}
\alias{coerce,character,CXErrorCode-method}
\alias{coerce,character,CXEvalResultKind-method}
\alias{coerce,character,CXGlobalOptFlags-method}
\alias{coerce,character,CXIdxAttrKind-method}
\alias{coerce,character,CXIdxDeclInfoFlags-method}
\alias{coerce,character,CXIdxEntityCXXTemplateKind-method}
\alias{coerce,character,CXIdxEntityKind-method}
\alias{coerce,character,CXIdxEntityLanguage-method}
\alias{coerce,character,CXIdxEntityRefKind-method}
\alias{coerce,character,CXIdxObjCContainerKind-method}
\alias{coerce,character,CXIndexOptFlags-method}
\alias{coerce,character,CXLanguageKind-method}
\alias{coerce,character,CXLinkageKind-method}
\alias{coerce,character,CXLoadDiag_Error-method}
\alias{coerce,character,CXNameRefFlags-method}
\alias{coerce,character,CXObjCDeclQualifierKind-method}
\alias{coerce,character,CXObjCPropertyAttrKind-method}
\alias{coerce,character,CXPrintingPolicyProperty-method}
\alias{coerce,character,CXRefQualifierKind-method}
\alias{coerce,character,CXReparse_Flags-method}
\alias{coerce,character,CXResult-method}
\alias{coerce,character,CXSaveError-method}
\alias{coerce,character,CXSaveTranslationUnit_Flags-method}
\alias{coerce,character,CXSymbolRole-method}
\alias{coerce,character,CXTLSKind-method}
\alias{coerce,character,CXTUResourceUsageKind-method}
\alias{coerce,character,CXTemplateArgumentKind-method}
\alias{coerce,character,CXTokenKind-method}
\alias{coerce,character,CXTranslationUnit_Flags-method}
\alias{coerce,character,CXTypeKind-method}
\alias{coerce,character,CXTypeLayoutError-method}
\alias{coerce,character,CXTypeNullabilityKind-method}
\alias{coerce,character,CXUnaryOperatorKind-method}
\alias{coerce,character,CXVisibilityKind-method}
\alias{coerce,character,CXVisitorResult-method}
\alias{coerce,character,CX_CXXAccessSpecifier-method}
\alias{coerce,character,CX_StorageClass-method}
\alias{coerce,integer,CXAvailabilityKind-method}
\alias{coerce,integer,CXBinaryOperatorKind-method}
\alias{coerce,integer,CXCallingConv-method}
\alias{coerce,integer,CXChildVisitResult-method}
\alias{coerce,integer,CXChoice-method}
\alias{coerce,integer,CXCodeComplete_Flags-method}
\alias{coerce,integer,CXCompletionChunkKind-method}
\alias{coerce,integer,CXCompletionContext-method}
\alias{coerce,integer,CXCursorKind-method}
\alias{coerce,integer,CXCursor_ExceptionSpecificationKind-method}
\alias{coerce,integer,CXDiagnosticDisplayOptions-method}
\alias{coerce,integer,CXDiagnosticSeverity-method}
\alias{coerce,integer,CXErrorCode-method}
\alias{coerce,integer,CXEvalResultKind-method}
\alias{coerce,integer,CXGlobalOptFlags-method}
\alias{coerce,integer,CXIdxAttrKind-method}
\alias{coerce,integer,CXIdxDeclInfoFlags-method}
\alias{coerce,integer,CXIdxEntityCXXTemplateKind-method}
\alias{coerce,integer,CXIdxEntityKind-method}
\alias{coerce,integer,CXIdxEntityLanguage-method}
\alias{coerce,integer,CXIdxEntityRefKind-method}
\alias{coerce,integer,CXIdxObjCContainerKind-method}
\alias{coerce,integer,CXIndexOptFlags-method}
\alias{coerce,integer,CXLanguageKind-method}
\alias{coerce,integer,CXLinkageKind-method}
\alias{coerce,integer,CXLoadDiag_Error-method}
\alias{coerce,integer,CXNameRefFlags-method}
\alias{coerce,integer,CXObjCDeclQualifierKind-method}
\alias{coerce,integer,CXObjCPropertyAttrKind-method}
\alias{coerce,integer,CXPrintingPolicyProperty-method}
\alias{coerce,integer,CXRefQualifierKind-method}
\alias{coerce,integer,CXReparse_Flags-method}
\alias{coerce,integer,CXResult-method}
\alias{coerce,integer,CXSaveError-method}
\alias{coerce,integer,CXSaveTranslationUnit_Flags-method}
\alias{coerce,integer,CXSymbolRole-method}
\alias{coerce,integer,CXTLSKind-method}
\alias{coerce,integer,CXTUResourceUsageKind-method}
\alias{coerce,integer,CXTemplateArgumentKind-method}
\alias{coerce,integer,CXTokenKind-method}
\alias{coerce,integer,CXTranslationUnit_Flags-method}
\alias{coerce,integer,CXTypeKind-method}
\alias{coerce,integer,CXTypeLayoutError-method}
\alias{coerce,integer,CXTypeNullabilityKind-method}
\alias{coerce,integer,CXUnaryOperatorKind-method}
\alias{coerce,integer,CXVisibilityKind-method}
\alias{coerce,integer,CXVisitorResult-method}
\alias{coerce,integer,CX_CXXAccessSpecifier-method}
\alias{coerce,integer,CX_StorageClass-method}
\alias{coerce,numeric,CXAvailabilityKind-method}
\alias{coerce,numeric,CXBinaryOperatorKind-method}
\alias{coerce,numeric,CXCallingConv-method}
\alias{coerce,numeric,CXChildVisitResult-method}
\alias{coerce,numeric,CXChoice-method}
\alias{coerce,numeric,CXCodeComplete_Flags-method}
\alias{coerce,numeric,CXCompletionChunkKind-method}
\alias{coerce,numeric,CXCompletionContext-method}
\alias{coerce,numeric,CXCursorKind-method}
\alias{coerce-numeric,CXCursor_ExceptionSpecificationKind-method}
\alias{coerce,numeric,CXDiagnosticDisplayOptions-method}
\alias{coerce,numeric,CXDiagnosticSeverity-method}
\alias{coerce,numeric,CXErrorCode-method}
\alias{coerce,numeric,CXEvalResultKind-method}
\alias{coerce,numeric,CXGlobalOptFlags-method}
\alias{coerce,numeric,CXIdxAttrKind-method}
\alias{coerce,numeric,CXIdxDeclInfoFlags-method}
\alias{coerce,numeric,CXIdxEntityCXXTemplateKind-method}
\alias{coerce,numeric,CXIdxEntityKind-method}
\alias{coerce,numeric,CXIdxEntityLanguage-method}
\alias{coerce,numeric,CXIdxEntityRefKind-method}
\alias{coerce,numeric,CXIdxObjCContainerKind-method}
\alias{coerce,numeric,CXIndexOptFlags-method}
\alias{coerce,numeric,CXLanguageKind-method}
\alias{coerce,numeric,CXLinkageKind-method}
\alias{coerce,numeric,CXLoadDiag_Error-method}
\alias{coerce,numeric,CXNameRefFlags-method}
\alias{coerce,numeric,CXObjCDeclQualifierKind-method}
\alias{coerce,numeric,CXObjCPropertyAttrKind-method}
\alias{coerce,numeric,CXPrintingPolicyProperty-method}
\alias{coerce,numeric,CXRefQualifierKind-method}
\alias{coerce,numeric,CXReparse_Flags-method}
\alias{coerce,numeric,CXResult-method}
\alias{coerce,numeric,CXSaveError-method}
\alias{coerce,numeric,CXSaveTranslationUnit_Flags-method}
\alias{coerce,numeric,CXSymbolRole-method}
\alias{coerce,numeric,CXTLSKind-method}
\alias{coerce,numeric,CXTUResourceUsageKind-method}
\alias{coerce,numeric,CXTemplateArgumentKind-method}
\alias{coerce,numeric,CXTokenKind-method}
\alias{coerce,numeric,CXTranslationUnit_Flags-method}
\alias{coerce,numeric,CXTypeKind-method}
\alias{coerce,numeric,CXTypeLayoutError-method}
\alias{coerce,numeric,CXTypeNullabilityKind-method}
\alias{coerce,numeric,CXUnaryOperatorKind-method}
\alias{coerce,numeric,CXVisibilityKind-method}
\alias{coerce,numeric,CXVisitorResult-method}
\alias{coerce,numeric,CX_CXXAccessSpecifier-method}
\alias{coerce,numeric,CX_StorageClass-method}
\title{S4 coerce methods}
\description{
  S4 methods for as()/coerce() to convert numbers to the corresponding
  enum value.
}


