\name{CXCallingConv}
\alias{CXCallingConv-class}
\docType{data}
\title{Enumeration type for calling convention}
%\description{}
%\usage{}
\format{
  The format is:
 Named int [1:12] 0 1 2 3 4 5 6 7 8 9 ...
 - attr(*, "names")= chr [1:12] "CXCallingConv_Default" "CXCallingConv_C" "CXCallingConv_X86StdCall" "CXCallingConv_X86FastCall" ...
}

\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
%\examples{}
\keyword{datasets}
