\name{CXCursor_ExceptionSpecificationKind}
\alias{CXCursor_ExceptionSpecificationKind-class}
\alias{coerce,numeric,CXCursor_ExceptionSpecificationKind-method}
\title{S4 class and method(s) for CXCursor_ExceptionSpecificationKind}
\description{
  S4 class and method for libclang's CXCursor_ExceptionSpecificationKind
  information.
}
