\name{CXCursorKind}
\alias{CXCursorKind}
\alias{CXCursor_UnexposedDecl}
\alias{CXCursor_StructDecl}
\alias{CXCursor_UnionDecl}
\alias{CXCursor_ClassDecl}
\alias{CXCursor_EnumDecl}
\alias{CXCursor_FieldDecl}
\alias{CXCursor_EnumConstantDecl}
\alias{CXCursor_FunctionDecl}
\alias{CXCursor_VarDecl}
\alias{CXCursor_ParmDecl}
\alias{CXCursor_ObjCInterfaceDecl}
\alias{CXCursor_ObjCCategoryDecl}
\alias{CXCursor_ObjCProtocolDecl}
\alias{CXCursor_ObjCPropertyDecl}
\alias{CXCursor_ObjCIvarDecl}
\alias{CXCursor_ObjCInstanceMethodDecl}
\alias{CXCursor_ObjCClassMethodDecl}
\alias{CXCursor_ObjCImplementationDecl}
\alias{CXCursor_ObjCCategoryImplDecl}
\alias{CXCursor_TypedefDecl}
\alias{CXCursor_CXXMethod}
\alias{CXCursor_Namespace}
\alias{CXCursor_LinkageSpec}
\alias{CXCursor_Constructor}
\alias{CXCursor_Destructor}
\alias{CXCursor_ConversionFunction}
\alias{CXCursor_TemplateTypeParameter}
\alias{CXCursor_NonTypeTemplateParameter}
\alias{CXCursor_TemplateTemplateParameter}
\alias{CXCursor_FunctionTemplate}
\alias{CXCursor_ClassTemplate}
\alias{CXCursor_ClassTemplatePartialSpecialization}
\alias{CXCursor_NamespaceAlias}
\alias{CXCursor_UsingDirective}
\alias{CXCursor_UsingDeclaration}
\alias{CXCursor_TypeAliasDecl}
\alias{CXCursor_ObjCSynthesizeDecl}
\alias{CXCursor_ObjCDynamicDecl}
\alias{CXCursor_CXXAccessSpecifier}
\alias{CXCursor_FirstDecl}
\alias{CXCursor_LastDecl}
\alias{CXCursor_FirstRef}
\alias{CXCursor_ObjCSuperClassRef}
\alias{CXCursor_ObjCProtocolRef}
\alias{CXCursor_ObjCClassRef}
\alias{CXCursor_TypeRef}
\alias{CXCursor_CXXBaseSpecifier}
\alias{CXCursor_TemplateRef}
\alias{CXCursor_NamespaceRef}
\alias{CXCursor_MemberRef}
\alias{CXCursor_LabelRef}
\alias{CXCursor_OverloadedDeclRef}
\alias{CXCursor_VariableRef}
\alias{CXCursor_LastRef}
\alias{CXCursor_FirstInvalid}
\alias{CXCursor_InvalidFile}
\alias{CXCursor_NoDeclFound}
\alias{CXCursor_NotImplemented}
\alias{CXCursor_InvalidCode}
\alias{CXCursor_LastInvalid}
\alias{CXCursor_FirstExpr}
\alias{CXCursor_UnexposedExpr}
\alias{CXCursor_DeclRefExpr}
\alias{CXCursor_MemberRefExpr}
\alias{CXCursor_CallExpr}
\alias{CXCursor_ObjCMessageExpr}
\alias{CXCursor_BlockExpr}
\alias{CXCursor_IntegerLiteral}
\alias{CXCursor_FloatingLiteral}
\alias{CXCursor_ImaginaryLiteral}
\alias{CXCursor_StringLiteral}
\alias{CXCursor_CharacterLiteral}
\alias{CXCursor_ParenExpr}
\alias{CXCursor_UnaryOperator}
\alias{CXCursor_ArraySubscriptExpr}
\alias{CXCursor_BinaryOperator}
\alias{CXCursor_CompoundAssignOperator}
\alias{CXCursor_ConditionalOperator}
\alias{CXCursor_CStyleCastExpr}
\alias{CXCursor_CompoundLiteralExpr}
\alias{CXCursor_InitListExpr}
\alias{CXCursor_AddrLabelExpr}
\alias{CXCursor_StmtExpr}
\alias{CXCursor_GenericSelectionExpr}
\alias{CXCursor_GNUNullExpr}
\alias{CXCursor_CXXStaticCastExpr}
\alias{CXCursor_CXXDynamicCastExpr}
\alias{CXCursor_CXXReinterpretCastExpr}
\alias{CXCursor_CXXConstCastExpr}
\alias{CXCursor_CXXFunctionalCastExpr}
\alias{CXCursor_CXXTypeidExpr}
\alias{CXCursor_CXXBoolLiteralExpr}
\alias{CXCursor_CXXNullPtrLiteralExpr}
\alias{CXCursor_CXXThisExpr}
\alias{CXCursor_CXXThrowExpr}
\alias{CXCursor_CXXNewExpr}
\alias{CXCursor_CXXDeleteExpr}
\alias{CXCursor_UnaryExpr}
\alias{CXCursor_ObjCStringLiteral}
\alias{CXCursor_ObjCEncodeExpr}
\alias{CXCursor_ObjCSelectorExpr}
\alias{CXCursor_ObjCProtocolExpr}
\alias{CXCursor_ObjCBridgedCastExpr}
\alias{CXCursor_PackExpansionExpr}
\alias{CXCursor_SizeOfPackExpr}
\alias{CXCursor_LambdaExpr}
\alias{CXCursor_ObjCBoolLiteralExpr}
\alias{CXCursor_LastExpr}
\alias{CXCursor_FirstStmt}
\alias{CXCursor_UnexposedStmt}
\alias{CXCursor_LabelStmt}
\alias{CXCursor_CompoundStmt}
\alias{CXCursor_CaseStmt}
\alias{CXCursor_DefaultStmt}
\alias{CXCursor_IfStmt}
\alias{CXCursor_SwitchStmt}
\alias{CXCursor_WhileStmt}
\alias{CXCursor_DoStmt}
\alias{CXCursor_ForStmt}
\alias{CXCursor_GotoStmt}
\alias{CXCursor_IndirectGotoStmt}
\alias{CXCursor_ContinueStmt}
\alias{CXCursor_BreakStmt}
\alias{CXCursor_ReturnStmt}
\alias{CXCursor_GCCAsmStmt}
\alias{CXCursor_AsmStmt}
\alias{CXCursor_ObjCAtTryStmt}
\alias{CXCursor_ObjCAtCatchStmt}
\alias{CXCursor_ObjCAtFinallyStmt}
\alias{CXCursor_ObjCAtThrowStmt}
\alias{CXCursor_ObjCAtSynchronizedStmt}
\alias{CXCursor_ObjCAutoreleasePoolStmt}
\alias{CXCursor_ObjCForCollectionStmt}
\alias{CXCursor_CXXCatchStmt}
\alias{CXCursor_CXXTryStmt}
\alias{CXCursor_CXXForRangeStmt}
\alias{CXCursor_SEHTryStmt}
\alias{CXCursor_SEHExceptStmt}
\alias{CXCursor_SEHFinallyStmt}
\alias{CXCursor_MSAsmStmt}
\alias{CXCursor_NullStmt}
\alias{CXCursor_DeclStmt}
\alias{CXCursor_LastStmt}
\alias{CXCursor_TranslationUnit}
\alias{CXCursor_FirstAttr}
\alias{CXCursor_UnexposedAttr}
\alias{CXCursor_IBActionAttr}
\alias{CXCursor_IBOutletAttr}
\alias{CXCursor_IBOutletCollectionAttr}
\alias{CXCursor_CXXFinalAttr}
\alias{CXCursor_CXXOverrideAttr}
\alias{CXCursor_AnnotateAttr}
\alias{CXCursor_AsmLabelAttr}
\alias{CXCursor_LastAttr}
\alias{CXCursor_PreprocessingDirective}
\alias{CXCursor_MacroDefinition}
\alias{CXCursor_MacroExpansion}
\alias{CXCursor_MacroInstantiation}
\alias{CXCursor_InclusionDirective}
\alias{CXCursor_FirstPreprocessing}
\alias{CXCursor_LastPreprocessing}
\alias{CXCursor_ModuleImportDecl}
\alias{CXCursor_FirstExtraDecl}
\alias{CXCursor_LastExtraDecl}
\alias{CXCursorKind-class}
 
\docType{data}
\title{Enumerations for different types of cursors/nodes in C parser tree}
\description{
  These are enumeration values identifying the types of the different
  cursors.
  One uses \code{cursor$kind} to get the cursor type of a particular
  cursor and then compare that value to any of these
  values to determine if the cursor is of that type.
}
\usage{CXCursorKind}
\format{
  The format is:
 Named int [1:3] 1 2 4
 - attr(*, "names")= chr [1:3] "CXDiagnostic_DisplaySourceLocation" "CXDiagnostic_DisplayColumn" "CXDiagnostic_DisplaySourceRanges"
}

\references{
   libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\examples{
CXCursorKind
}
\keyword{datasets}
