\name{CXEnums}
\alias{CXCursor_AlignedAttr}
\alias{CXCursor_BuiltinBitCastExpr}
\alias{CXCursor_ConceptDecl}
\alias{CXCursor_ConceptSpecializationExpr}
\alias{CXCursor_ConvergentAttr}
\alias{CXCursor_CUDAConstantAttr}
\alias{CXCursor_CUDADeviceAttr}
\alias{CXCursor_CUDAGlobalAttr}
\alias{CXCursor_CUDAHostAttr}
\alias{CXCursor_CUDASharedAttr}
\alias{CXCursor_CXXAddrspaceCastExpr}
\alias{CXCursor_CXXParenListInitExpr}
\alias{CXCursor_DLLExport}
\alias{CXCursor_DLLImport}
\alias{CXCursor_ExceptionSpecificationKind}
\alias{CXCursor_ExceptionSpecificationKind_BasicNoexcept}
\alias{CXCursor_ExceptionSpecificationKind_ComputedNoexcept}
\alias{CXCursor_ExceptionSpecificationKind_Dynamic}
\alias{CXCursor_ExceptionSpecificationKind_DynamicNone}
\alias{CXCursor_ExceptionSpecificationKind_MSAny}
\alias{CXCursor_ExceptionSpecificationKind_None}
\alias{CXCursor_ExceptionSpecificationKind_NoThrow}
\alias{CXCursor_ExceptionSpecificationKind_Unevaluated}
\alias{CXCursor_ExceptionSpecificationKind_Uninstantiated}
\alias{CXCursor_ExceptionSpecificationKind_Unparsed}
\alias{CXCursor_ExceptionSpecificationKindValues}
\alias{CXCursor_FixedPointLiteral}
\alias{CXCursor_FlagEnum}
\alias{CXCursor_FriendDecl}
\alias{CXCursor_NSConsumed}
\alias{CXCursor_NSConsumesSelf}
\alias{CXCursor_NSReturnsAutoreleased}
\alias{CXCursor_NSReturnsNotRetained}
\alias{CXCursor_NSReturnsRetained}
\alias{CXCursor_ObjCAvailabilityCheckExpr}
\alias{CXCursor_ObjCBoxable}
\alias{CXCursor_ObjCDesignatedInitializer}
\alias{CXCursor_ObjCException}
\alias{CXCursor_ObjCExplicitProtocolImpl}
\alias{CXCursor_ObjCIndependentClass}
\alias{CXCursor_ObjCNSObject}
\alias{CXCursor_ObjCPreciseLifetime}
\alias{CXCursor_ObjCRequiresSuper}
\alias{CXCursor_ObjCReturnsInnerPointer}
\alias{CXCursor_ObjCRootClass}
\alias{CXCursor_ObjCRuntimeVisible}
\alias{CXCursor_ObjCSubclassingRestricted}
\alias{CXCursor_OMPArraySectionExpr}
\alias{CXCursor_OMPArrayShapingExpr}
\alias{CXCursor_OMPAtomicDirective}
\alias{CXCursor_OMPBarrierDirective}
\alias{CXCursor_OMPCancelDirective}
\alias{CXCursor_OMPCancellationPointDirective}
\alias{CXCursor_OMPCanonicalLoop}
\alias{CXCursor_OMPCriticalDirective}
\alias{CXCursor_OMPDepobjDirective}
\alias{CXCursor_OMPDispatchDirective}
\alias{CXCursor_OMPDistributeDirective}
\alias{CXCursor_OMPDistributeParallelForDirective}
\alias{CXCursor_OMPDistributeParallelForSimdDirective}
\alias{CXCursor_OMPDistributeSimdDirective}
\alias{CXCursor_OMPErrorDirective}
\alias{CXCursor_OMPFlushDirective}
\alias{CXCursor_OMPForDirective}
\alias{CXCursor_OMPForSimdDirective}
\alias{CXCursor_OMPGenericLoopDirective}
\alias{CXCursor_OMPInteropDirective}
\alias{CXCursor_OMPIteratorExpr}
\alias{CXCursor_OMPMaskedDirective}
\alias{CXCursor_OMPMaskedTaskLoopDirective}
\alias{CXCursor_OMPMaskedTaskLoopSimdDirective}
\alias{CXCursor_OMPMasterDirective}
\alias{CXCursor_OMPMasterTaskLoopDirective}
\alias{CXCursor_OMPMasterTaskLoopSimdDirective}
\alias{CXCursor_OMPMetaDirective}
\alias{CXCursor_OMPOrderedDirective}
\alias{CXCursor_OMPParallelForDirective}
\alias{CXCursor_OMPParallelForSimdDirective}
\alias{CXCursor_OMPParallelGenericLoopDirective}
\alias{CXCursor_OMPParallelMaskedDirective}
\alias{CXCursor_OMPParallelMaskedTaskLoopDirective}
\alias{CXCursor_OMPParallelMaskedTaskLoopSimdDirective}
\alias{CXCursor_OMPParallelMasterDirective}
\alias{CXCursor_OMPParallelMasterTaskLoopDirective}
\alias{CXCursor_OMPParallelMasterTaskLoopSimdDirective}
\alias{CXCursor_OMPParallelSectionsDirective}
\alias{CXCursor_OMPScanDirective}
\alias{CXCursor_OMPScopeDirective}
\alias{CXCursor_OMPSectionDirective}
\alias{CXCursor_OMPSectionsDirective}
\alias{CXCursor_OMPSingleDirective}
\alias{CXCursor_OMPTargetDataDirective}
\alias{CXCursor_OMPTargetDirective}
\alias{CXCursor_OMPTargetEnterDataDirective}
\alias{CXCursor_OMPTargetExitDataDirective}
\alias{CXCursor_OMPTargetParallelDirective}
\alias{CXCursor_OMPTargetParallelForDirective}
\alias{CXCursor_OMPTargetParallelForSimdDirective}
\alias{CXCursor_OMPTargetParallelGenericLoopDirective}
\alias{CXCursor_OMPTargetSimdDirective}
\alias{CXCursor_OMPTargetTeamsDirective}
\alias{CXCursor_OMPTargetTeamsDistributeDirective}
\alias{CXCursor_OMPTargetTeamsDistributeParallelForDirective}
\alias{CXCursor_OMPTargetTeamsDistributeParallelForSimdDirective}
\alias{CXCursor_OMPTargetTeamsDistributeSimdDirective}
\alias{CXCursor_OMPTargetTeamsGenericLoopDirective}
\alias{CXCursor_OMPTargetUpdateDirective}
\alias{CXCursor_OMPTaskDirective}
\alias{CXCursor_OMPTaskgroupDirective}
\alias{CXCursor_OMPTaskLoopDirective}
\alias{CXCursor_OMPTaskLoopSimdDirective}
\alias{CXCursor_OMPTaskwaitDirective}
\alias{CXCursor_OMPTaskyieldDirective}
\alias{CXCursor_OMPTeamsDirective}
\alias{CXCursor_OMPTeamsDistributeDirective}
\alias{CXCursor_OMPTeamsDistributeParallelForDirective}
\alias{CXCursor_OMPTeamsDistributeParallelForSimdDirective}
\alias{CXCursor_OMPTeamsDistributeSimdDirective}
\alias{CXCursor_OMPTeamsGenericLoopDirective}
\alias{CXCursor_OMPTileDirective}
\alias{CXCursor_OMPUnrollDirective}
\alias{CXCursor_OverloadCandidate}
\alias{CXCursor_RequiresExpr}
\alias{CXCursor_SEHLeaveStmt}
\alias{CXCursor_StaticAssert}
\alias{CXCursor_TypeAliasTemplateDecl}
\alias{CXCursor_VisibilityAttr}
\alias{CXCursor_WarnUnusedAttr}
\alias{CXCursor_WarnUnusedResultAttr}
\alias{CXTranslationUnit_CacheCompletionResults}
\alias{CXTranslationUnit_CXXChainedPCH}
\alias{CXTranslationUnit_DetailedPreprocessingRecord}
\alias{CXTranslationUnit_ForSerialization}
\alias{CXTranslationUnit_IncludeBriefCommentsInCodeCompletion}
\alias{CXTranslationUnit_Incomplete}
\alias{CXTranslationUnit_None}
\alias{CXTranslationUnit_PrecompiledPreamble}
\alias{CXTranslationUnit_SkipFunctionBodies}
\title{CXCursor and CXTranslationUnit Enums}
\description{
  R-level values corresponding to many of libclang's C-level enum values.
}
