\name{enableStackTraces}
\alias{enableStackTraces}
\alias{toggleCrashRecovery}
\title{Set global state in libclang}
\description{
}
\usage{
enableStackTraces()
toggleCrashRecovery(isEnabled)
}
\arguments{
   \item{isEnabled}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
