\name{equalCursors}
\alias{equalCursors}
\alias{equalLocations}
\alias{equalRanges}
\alias{equalTypes}
\title{Functions to determine if two instances of a data type are the same}
\description{
}
\usage{
equalCursors(arg1, arg2)
equalLocations(loc1, loc2)
equalRanges(range1, range2)
equalTypes(A, B)
}
\arguments{
   \item{arg1}{}
   \item{arg2}{}
   \item{loc1}{}
   \item{loc2}{}
   \item{range1}{}
   \item{range2}{}
   \item{A}{}
   \item{B}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
