\name{getChildren}
\alias{getChildren}
\title{Collect the child cursors of a given cursor}
\description{
  This visits the immediate child cursors of the given cursor
  and collects them (cloning them).
}
\usage{
getChildren(x)
}
\arguments{
   \item{x}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
