\name{Module_getFullName}
\alias{Module_getFullName}
\alias{Module_getName}
\alias{Module_getNumTopLevelHeaders}
\alias{Module_getParent}
\alias{Module_getTopLevelHeader}
\alias{getModule}
\title{Functions for querying a libclang Module}
\description{
}
\usage{
getModule(C)
Module_getFullName(Module)
Module_getName(Module)
Module_getNumTopLevelHeaders(arg1, Module)
Module_getParent(Module)
Module_getTopLevelHeader(arg1, Module, Index)
}
\arguments{
  \item{C}{}
   \item{Module}{}
   \item{arg1}{}
   \item{Index}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
