\name{enableStackTraces}
\alias{enableStackTraces}
\alias{toggleCrashRecovery}
\title{Set global state in libclang}
\description{
  These are functions for setting global state about how libclang behaves.
}
\usage{
enableStackTraces()
toggleCrashRecovery(isEnabled)
}
\arguments{
   \item{isEnabled}{a logical value}
}
%\value{}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
%\seealso{}
%\examples{}
\keyword{programming}
\keyword{interface} 
