\name{findCalls}
\alias{findCalls}
\title{Get the names of all routines called in any code in a translation
unit}
\description{
}
\usage{
findCalls(file, n = 10000, ...)
}
\arguments{
   \item{file}{}
   \item{n}{}
   \item{...}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
