\name{getArraySize}
\alias{getArraySize}
\alias{getArrayElementType}
\alias{getNumElements}
\alias{getElementType}
\title{Get information about container types}
\description{
}
\usage{
getArraySize(T)
getArrayElementType(T)
getNumElements(T)
getElementType(T)
}
\arguments{
   \item{T}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
