\name{getCXTUResourceUsage}
\alias{getCXTUResourceUsage}
\alias{getTUResourceUsage}
\title{Information about the resources a translation unit consumed}
\description{
}
\usage{
getCXTUResourceUsage(tu)
}
\arguments{
   \item{tu}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
