\name{getDataStructures}
\alias{getDataStructures}
\alias{getStructDef}
\title{Get the top-level data types in a translation unit}
\description{
}
\usage{
getDataStructures(src, collector = genDataStructCollector(), ...)
getStructDef(type, kids = children(tp))
}
\arguments{
   \item{src}{}
   \item{collector}{}
   \item{...}{}
   \item{type}{}
   \item{kids}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
