\name{getDefines}
\alias{getDefines}
\title{Get preprocessor defines in a translation unit}
\description{
}
\usage{
getDefines(tu, names, col = genDefinesCollector(names))
}
\arguments{
   \item{tu}{}
   \item{names}{}
   \item{col}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
