\name{getEnums}
\alias{getEnums}
\title{Get the top-level enumeration definitions in a translation unit}
\description{
}
\usage{
getEnums(src)
}
\arguments{
   \item{src}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
