\name{getInclusions}
\alias{getInclusions}
\title{Collect all of the include files for a source file}
\description{
}
\usage{
getInclusions(file, fun = genInclusionCollector(), ...)
}
\arguments{
   \item{file}{}
   \item{fun}{}
   \item{...}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
