\name{getInstantiationLocation}
\alias{getInstantiationLocation}
\alias{getExpansionLocation}
\alias{getPresumedLocation}
\alias{getLocation}
\alias{getRange}
\alias{getNullLocation}
\alias{isInSystemHeader}
\alias{Location_isInSystemHeader}
\title{Functions involving Location information}
\description{
}
\usage{
getInstantiationLocation(x)
getRange(begin, end)
getNullLocation()
isInSystemHeader(location)
}
\arguments{
   \item{x}{}
   \item{begin}{}
   \item{end}{}
   \item{location}{}
}
\value{
}
\references{
 libclang \url{http://clang.llvm.org/doxygen/group__CINDEX.html}
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\keyword{interface} 
